raise 404 on delete action

+++ ckanext/webhooks/actions.py
@@ -54,6 +54,9 @@ def webhook_delete(context, data_dict):
         raise ValidationError(errors)
 
     webhook = db.Webhook.get(id=data['id'])
+    if webhook is None:
+        raise NotFound()
+        
     session = context['session']
     session.delete(webhook)
     session.commit()
